+++ jobmon_gui/src/components/workflow_details/MemoryHistogram.tsx
