+++ jobmon_gui/src/components/workflow_details/Tasks.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import TaskTable from './task_table';
+import TaskTable from './TaskTable';
 import { FaSearch } from "react-icons/fa";
 
 import { safe_rum_start_span, safe_rum_unit_end } from '../../utils/rum'
