+++ jobmon_gui/src/components/workflow_details/Usage.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
-import MemoryHistogram from './memory_histogram';
-import RuntimeHistogram from './runtime_histogram';
+import MemoryHistogram from './MemoryHistogram';
+import RuntimeHistogram from './RuntimeHistogram';
 import { formatBytes, bytes_to_gib } from '../../utils/formatters'
 import { safe_rum_start_span, safe_rum_unit_end } from '../../utils/rum'
 import humanizeDuration from 'humanize-duration';
