+++ jobmon_gui/src/components/workflow_details/WorkflowHeader.tsx
@@ -9,7 +9,7 @@ import { AiFillSchedule, AiFillCheckCircle} from "react-icons/ai";
 import { TbHandStop } from "react-icons/tb";
 import {HiRocketLaunch} from "react-icons/hi2";
 
-export default function WFHeader({
+export default function WorkflowHeader({
     wf_id,
     wf_status,
     wf_status_desc,
