+++ jobmon_gui/src/components/workflow_overview/WorkflowTable.tsx
@@ -12,7 +12,7 @@ import { convertDate, convertDatePST } from '../../utils/formatters';
 import { FaCaretDown, FaCaretUp, FaCircle } from "react-icons/fa";
 
 // @ts-ignore
-import JobmonProgressBar from '../../components/JobmonProgressBar.tsx';
+import JobmonProgressBar from '../JobmonProgressBar.tsx';
 
 
 const customCaret = (order, column) => {
@@ -81,7 +81,7 @@ interface Data {
     wfr_count: number
 }
 
-export default function JobmonWFTable({ allData }) {
+export default function WorkflowTable({ allData }) {
     const [expandedRows, setExpandedRows] = useState([]);
     const [statusDict, setStatusDict] = useState({});
     //TODO: get rid of finishedWF
