+++ jobmon_gui/src/screens/WorkflowDetails.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import '../../styles/jobmon_gui.css';
+import '../styles/jobmon_gui.css';
 // import '../../styles/jobmon_gui.css';
 import { useParams, Link, Outlet, useNavigate } from 'react-router-dom';
 import { useForm } from "react-hook-form";
@@ -12,13 +12,13 @@ import humanizeDuration from 'humanize-duration';
 
 
 // @ts-ignore
-import JobmonProgressBar from '../../components/JobmonProgressBar.tsx';
-import Tasks from './tasks';
-import Usage from './usage';
-import Errors from './errors';
-import WFHeader from "./wf_header"
-import { convertDatePST } from '../../utils/formatters';
-import { init_apm, safe_rum_add_label, safe_rum_transaction } from '../../utils/rum';
+import JobmonProgressBar from '../components/JobmonProgressBar.tsx';
+import Tasks from '../components/workflow_details/Tasks';
+import Usage from '../components/workflow_details/Usage';
+import Errors from '../components/workflow_details/Errors';
+import WorkflowHeader from "../components/workflow_details/WorkflowHeader"
+import { convertDatePST } from '../utils/formatters';
+import { init_apm, safe_rum_add_label, safe_rum_transaction } from '../utils/rum';
 
 function getAsyncWFdetail(setWFDict, wf_id: string) {
     const url = process.env.REACT_APP_BASE_URL + "/workflow_status_viz";
@@ -259,7 +259,7 @@ function WorkflowDetails({ subpage }) {
                 <Breadcrumb.Item active>Workflow ID {workflowId} </Breadcrumb.Item>
             </Breadcrumb>
             <div className='d-flex justify-content-start pt-3'>
-                <WFHeader
+                <WorkflowHeader
                       wf_id={workflowId}
                       wf_status={wf_status}
                       wf_status_desc={wf_status_desc}
